FitnessChatbot/requirements.txt
pandas
numpy
streamlit
google-generativeai
python-dotenvs